ui/requirements.txt
streamlit>=1.28.0
ultralytics
opencv-python
pillow
numpy
